don't set foldmethod
The defaults in Vim mean that folds will start out closed, if we set
foldmethod to something other than manual.  This is rather annoying for
people who don't like or use folding, and who leave the relevant options
at their defaults.  Thus, folding should be opt-in, not opt-out.  The
simplest way to achieve this is probably to leave it up to the user to
set foldmethod approprietly.  We do however set foldexpr since this
doesn't do anything until foldmethod is set to expr and advanced users
can easily override it should they want to.

Fixes #8 but needs more work in #10