fix spacing in for snippet
I had added the extra spacing as a work-around for a bug in UltiSnips
which turns out to be a result of a Vim option I had set to a
non-default value.  It's still a bug in UltiSnips, but most people will
not be affected and my work-around will do more harm than good.